Use numeric versions for the binary
The binary is updated less often than the Java part.
A version string suggests that the version number should be the same
as for the Java part (especially when the strings are similar), but
this is not the case.

The different format should avoid this misunderstanding and also
recompiling the binary just to change the version from "beta"
to "release" etc.