v44: Several performance improvements
The main change is that the original method is now saved in insns.
This does not only save a lookup, but we can also save additional
information with it. We save the Method object (to avoid creating a
new one every time) and an arbitrary object (which is filled with
references to callbacks, return type and parameter types by XposedBridge.

We also do the lookup of the Object class just once.

Requires changes in XposedBridge to work.