Add key triggers for safe mode during boot.
To disable Xposed during zygote startup:
- Make sure one of the VOLUME keys is held down before app_process starts
- 2 short vibrations will occur when the detection starts (zygote starting)
- press and release POWER or HOME twice, without releasing the VOLUME key that is held down
- release the VOLUME key
- 2 short vibrations will signal that the combo was detected and the disabler file will be created

To enable Xposed, the same steps can be executed but with three presses of POWER/HOME instead of two.
Three vibrations will occur if this was triggered.

If VOLUME was pressed but no valid combination occurs within 10 seconds, a single vibration will
signal that an invalid combination was used.