New logic for safemode detection, based on a single key being pressed multiple times.
Upon startup, one of the physical keys must be pressed during the first 2 seconds (or already
held down upfront).
After that, the same key must be pressed again within 5 seconds, until 5 total presses are reached.

Two short vibrations are issued when the initial key is detected, and an additional short vibration
for every subsequent key press.
If the trigger is successful, a long vibration is issued.