v55: Don't load Xposed for commandline tools
Examples for such tools are "am" and "pm".

The IXposedHookCmdInit / initCmdApp() API was only used by two modules in the
in the repository, and both of them actually didn't need it (and therefore
removed it now). So let's not execute our code unnecessarily. It also reduces
the logging, which is good because log lines of Zygote and tools could mix up
and confuse users.

It's still possible to enable processing by creating a file, whoever no UI
option is planned. It's mainly for testing very low-level changes in Xposed.