v57: Make offset for the JIT reset flag configurable
Some ROMs seem to be compiled with different settings or have been
modified by the vendors, resulting in a different offset for the
codeCacheFull flag than 0x78 = 120 as in AOSP. This can lead to
boot loops: https://github.com/rovo89/Xposed/issues/6

Manual configuration is only a workaround, however it's better than
compiling different variants. Automated detection would be best,
however it's very complicated. Not only does it require intelligent
algorithms to guess the offset from other, more significant fields,
it would also have to be done at the right time. The structure in
question isn't initialized in Zygote, and only after some delay
when starting apps.

The only thing we can do now is a quick sanity check. Booleans can
only be 0 or 1, so if the current value is something else, then we
were probably about to overwrite a pointer or something else.