Lots of changes and refactoring for Lollipop
Sorry for the huge commit. The code has evolved over a period of more than
one year, with lots of changes and several rounds of refactoring.

The changes include:
- Move Dalvik-specific code to a separate library.
- Create a similar library for ART (although much of the code is actually
  in the modified ART runtime).
- Load the runtime-specific libary only when required.
- Implement services for accessing files to work around SELinux restrictions.
- Implement a daemon that writes certain logcat lines to a file.
- Other adjustments for Lollipop and refactoring.

Generally, this is supposed to work on older Android versions as well, but
it hasn't been tested very well so far.