Don't use writeByteArray() in xposed_service.cpp
I'm not sure the reasoning behind using this on Lollipop+
and writeInt32()/write() on older versions.  The non-LP
code seems to work fine on LP, and using writeByteArray()
breaks 64-bit builds.

writeByteArray() breaks 64-bit builds because it writes
the size of the byte array first.  Normally, this is not
breaking.  But because the C++ implementation for
writeByteArray() uses a size_t and we only pass a 32-bit
integer, our value is padded.  When we later go back to
read this data, we only read a 32-bit integer for the length
of our byte array.  On 32-bit this works fine, on 64-bit
we now have 32-bit's of zeroes unaccounted for.  This means
on 64-bit our file reads start "4 bytes early" and finish
"4 bytes early".  This is most pronounced on XSharedPreferences
where we end up reading invalid XML.

This fixes rovo89/XposedBridge#58.

Thanks to @rsteckler for reporting this issue and testing the fix!