Handle different variants of AndroidRuntime::start()
In Android 5.1.1_r19, Google introduced an additonal parameter for this
method. We need to detect which one the ROM uses and call it.

This assumes that the mangled name is consistent across all ROMs and
that the object instance is implicitely handled as first parameter
to the method. On Android 5.0, we assume that this change is not
merged. If required, the dynamic way could simply be enabled for this
version as well.

Fixes #71.