Fix invokeOriginalMethod()
In most cases, invokeOriginalMethodNative() is called from
handleHookedMethod() with the already resolved original method.

However, it's also allowed to call it via invokeOriginalMethod().
In that case, it's possible that the method is actually the hooked one.
Make sure to resolve it to the original one if the second parameter is
zero. The ART variant will be changed to set it to 1 before calling
handleHookedMethod().