Adjust SYSTEMSERVERCLASSPATH for HTC devices
They make use of a HtcDeviceInfoManager class from /system/framework/ub.jar
in the ActivityManagerService.  This works fine on odex'ed ROMs because
the class is precompiled. But with Xposed, the odex file can't be used
as-is and has to be recompiled before. This would be done later by the
PackageManagerService, but then it's too late.

As a solution, add ub.jar to the SYSTEMSERVERCLASSPATH if it exists.
In the end, it's exactly that: A class which is used by the system server.