Don't include generic crash logs by default
The crashes can be caused by anything, not just Xposed or ART.
This sometimes leads to confusion, and the crash log itself is often
not even helpful because other relevant information isn't included.

Now that it's possible to enable a full logcat, don't include those
crashes in the "normal" logs.